Minor refactor - Don’t expose ReadInfo and WriteInfo - Rename AutoValueExtensionValidator to PaperParcelAutoValueExtensionValidator
